Change some minor stuffs

+++ src/que2.c
@@ -3,22 +3,22 @@
 
 void que2_init(que2 queue)
 {
-    queue->max_size = 4 * sizeof(size_t) * 8;
-    queue->data = (size_t*) malloc_exit_when_null(queue->max_size * sizeof(size_t));
+    queue->max_size = 4 * sizeof(signed long) * 8;
+    queue->data = (signed long*) malloc_exit_when_null(queue->max_size * sizeof(signed long));
     queue->size = 0;
 }
 
-void que2_push(que2 queue, size_t val1, size_t val2)
+void que2_push(que2 queue, signed long val1, signed long val2)
 {
     if (queue->max_size > queue->size + 2) {
-        queue->max_size += 4 * sizeof(size_t) * 8;
-        queue->data = (size_t*) realloc(queue->data, queue->max_size * sizeof(size_t));
+        queue->max_size += 4 * sizeof(signed long) * 8;
+        queue->data = (signed long*) realloc(queue->data, queue->max_size * sizeof(signed long));
     }
     queue->data[queue->size++] = val1;
     queue->data[queue->size++] = val2;
 }
 
-int que2_pop(que2 queue, size_t* val1, size_t* val2) 
+int que2_pop(que2 queue, signed long* val1, signed long* val2) 
 {
     if (queue->size < 2)
         return 0;
